Fix được lỗi load database

+++ app/src/main/java/com/example/tina/doanmang_tinakeeper/model/Expense.java
@@ -77,7 +77,7 @@ public class Expense {
         return money;
     }
 
-    public void setMoney(int money) {
+    public void setMoney(double money) {
         this.money = money;
     }
